+++ src/main/java/org/webbitserver/WebSocketHandler.java
@@ -1,7 +1,7 @@
 package org.webbitserver;
 
 public interface WebSocketHandler {
-    void onOpen(WebSocketConnection connection) throws Exception;
+    void onOpen(WebSocketConnection connection) throws Throwable;
 
     /**
      * Called when a connection is closed.
@@ -9,7 +9,7 @@ public interface WebSocketHandler {
      * @param connection the connection that was closed. Beware that the connection will be null if this handler is used in a {@link WebSocket} that fails to connect.
      * @throws Exception
      */
-    void onClose(WebSocketConnection connection) throws Exception;
+    void onClose(WebSocketConnection connection) throws Throwable;
 
     void onMessage(WebSocketConnection connection, String msg) throws Throwable;